+++ CHANGELOG.md
@@ -21,11 +21,12 @@ Clean up and Stability patch.
 ### Added
 
 - Increased unit test coverage.
-- Improved examples on homepage: live preview, mini-ide, diff?, performance 
+- Improved examples on homepage: live preview, theme, performance, ide?
 
 ### Changed
 
 - Visible range of editor to be more consistent.
+- Theme styling to use color and background color for scrollbars.
 
 ## [1.1.0] - 2023-07-29